Authorization added

+++ core/Portfi.Infrastructure/Services/Implementations/PortfolioService.cs
@@ -20,6 +20,7 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> AddProjectsByPortfolioId(
         string portfolioId,
+        string personId,
         string[] sourceCodeLinks)
     {
         var foundPortfolio = await portfolioRepository
@@ -27,6 +28,11 @@ public class PortfolioService(
             .Include(p => p.Projects)
             .FirstOrDefaultAsync(p => p.Id == Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         foreach (string sourceCodeLink in sourceCodeLinks)
         {
@@ -51,6 +57,7 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> AddSocialMediaLinks(
         string portfolioId,
+        string personId,
         string serializedSocialMediaLinks)
     {
         IEnumerable<REQUESTS.AddSocialMediaLinkRequest> socialMediaLinkRequests = [];
@@ -83,6 +90,10 @@ public class PortfolioService(
             .FirstOrDefaultAsync(p => p.Id == Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
 
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         foreach (REQUESTS.AddSocialMediaLinkRequest socialMediaLink in socialMediaLinkRequests)
         {
@@ -135,10 +146,16 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> EditBiographyByPortfolioId(
         string portfolioId,
+        string personId,
         string newBiography)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         if (foundPortfolio.Biography != newBiography)
         {
@@ -156,10 +173,16 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> EditNamesByPortfolioId(
         string portfolioId,
+        string personId,
         string[] newNames)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         foundPortfolio.PersonNames = newNames;
 
@@ -174,12 +197,21 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> EditSocialMediaLinkById(
         string socialMediaLinkId,
+        string personId,
         string newSocialMediaLink)
     {
         var foundSocialMediaLink = await socialMediaLinkRepository.GetByIdAsync(Guid.Parse(socialMediaLinkId))
             ?? throw new ArgumentNullException("Social media link not found.");
 
         Guid portfolioId = foundSocialMediaLink.PortfolioId;
+
+        var foundPortfolio = await portfolioRepository.GetByIdAsync(portfolioId)
+            ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         if (foundSocialMediaLink.Value != newSocialMediaLink)
         {
@@ -191,20 +223,23 @@ public class PortfolioService(
             }
         }
 
-        var foundPortfolio = await portfolioRepository.GetByIdAsync(portfolioId)
-            ?? throw new ArgumentNullException("Portfolio not found.");
-
         return foundPortfolio;
     }
 
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> EditThemeByPortfolioId(
         string portfolioId,
+        string personId,
         string? newBackgroundTheme,
         string? newMainColor)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         if (!string.IsNullOrEmpty(newBackgroundTheme)
             && foundPortfolio.BackgroundTheme != newBackgroundTheme)
@@ -229,10 +264,16 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> EditVisabilityByPortfolioId(
         string portfolioId,
+        string personId,
         bool isPublic)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         if (foundPortfolio.IsPublic != isPublic)
         {
@@ -275,10 +316,16 @@ public class PortfolioService(
 
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> RemoveAvatarByPortfolioId(
-        string portfolioId)
+        string portfolioId,
+        string personId)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         foundPortfolio.Avatar = PCONST.DefaultAvatarValue;
 
@@ -292,10 +339,16 @@ public class PortfolioService(
 
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> RemoveCVByPortfolioId(
-        string portfolioId)
+        string portfolioId,
+        string personId)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         foundPortfolio.CV = null;
 
@@ -310,30 +363,42 @@ public class PortfolioService(
 
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> RemoveSocialMediaLinkById(
-        string socialMediaLinkId)
+        string socialMediaLinkId,
+        string personId)
     {
         var foundSocialMediaLink = await socialMediaLinkRepository.GetByIdAsync(Guid.Parse(socialMediaLinkId))
             ?? throw new ArgumentNullException("Social media link not found.");
 
         Guid portfolioId = foundSocialMediaLink.PortfolioId;
 
+        var foundPortfolio = await portfolioRepository.GetByIdAsync(portfolioId)
+            ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
+
         if (!await socialMediaLinkRepository.DeleteAsync(foundSocialMediaLink))
         {
             throw new EXCEPTIONS.ItemNotDeletedException($"{nameof(foundSocialMediaLink)} cannot be deleted.");
         }
-
-        var foundPortfolio = await portfolioRepository.GetByIdAsync(portfolioId)
-            ?? throw new ArgumentNullException("Portfolio not found.");
 
         return foundPortfolio;
     }
 
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> SetDefaultThemeByportfolioId(
-        string portfolioId)
+        string portfolioId,
+        string personId)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         bool isUpdated = false;
 
@@ -365,10 +430,16 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> UplaodAvatarByPortfolioId(
         string portfolioId,
+        string personId,
         string avatarURL)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         if (foundPortfolio.Avatar != avatarURL)
         {
@@ -386,10 +457,16 @@ public class PortfolioService(
     /// <inheritdoc/>
     async public Task<MODELS.Portfolio> UplaodCVByPortfolioId(
         string portfolioId,
+        string personId,
         string cvURL)
     {
         var foundPortfolio = await portfolioRepository.GetByIdAsync(Guid.Parse(portfolioId))
             ?? throw new ArgumentNullException("Portfolio not found.");
+
+        if (foundPortfolio.PersonId != personId)
+        {
+            throw new EXCEPTIONS.NotAuthorizedException($"No permission for user with ID `{personId}`.");
+        }
 
         if (foundPortfolio.CV != cvURL)
         {
